feature_1/requirements.txt
Django
django-import-export
django-filter
django-crispy-forms
django-widget-tweaks
